feat: refactor lint, get chat fix

+++ internal/infrastructure/sql/storages/commands_medias/models/models.go
@@ -11,7 +11,7 @@ type CommandMedia struct {
 	Name      string     `db:"name"       json:"name"`
 	Type      string     `db:"type"       json:"type"`
 	URL       string     `db:"url"        json:"url"`
-	Size      int        `db:"size"       json:"size"`
+	Size      uint64     `db:"size"       json:"size"`
 	CommandID *uuid.UUID `db:"command_id" json:"command_id,omitempty"`
 	UserID    *uuid.UUID `db:"user_id"    json:"user_id,omitempty"`
 	CreatedAt time.Time  `db:"created_at" json:"created_at"`
